s21-day06/wip -> ex01 : add img/awesome_logo.png, start work with database
