+++ day06/ex01/internal/database/postgres.go
@@ -1,0 +1,1 @@
+package database
